+++ lib/ts/recipe/thirdparty/utils.ts
@@ -29,7 +29,6 @@ import {
     TypeInputSessionFeature,
     TypeNormalisedInputSessionFeature,
 } from "./types";
-import { DEV_OAUTH_CLIENT_IDS } from "./constants";
 
 export function validateAndNormaliseUserInput(
     recipeInstance: Recipe,
@@ -144,8 +143,4 @@ function validateAndNormaliseEmailVerificationConfig(
                       return await config.emailVerificationFeature.getEmailVerificationURL(userInfo);
                   },
     };
-}
-
-export function isUsingOAuthDevelopmentKeys(client_id: string): boolean {
-    return DEV_OAUTH_CLIENT_IDS.includes(client_id);
 }